fix the six pre-existing lints the clippy gate surfaced
the gate's first full pass earned its keep: map_identity ×2 and
useless_conversion in agent/http.rs, unnecessary_cast in tools.rs'
civil_from_days (hinnant arithmetic already yields i64), redundant_closure
in ui/mod.rs' on_click (Closure::new takes the fn directly), and
derivable_impls for WorkerState::default — which now also documents that
every field's zero value is the true initial state. no behavior change by
construction: every fix is what clippy itself suggested, and the 71-test
suite passed on the previous attempt before clippy ran.

+++ src/agent/tools.rs
@@ -99,7 +99,7 @@ pub fn now_epoch_ms() -> i64 {
 fn civil_from_days(days: i64) -> (i64, u32, u32) {
     let z = days + 719_468;
     let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = (z - era * 146_097) as i64; // [0, 146096]
+    let doe = z - era * 146_097; // [0, 146096]
     let yoe = (doe - doe / 1460 + doe / 36_524 - doe / 146_096) / 365; // [0, 399]
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100); // [0, 365]
